Add basic authentication

+++ My Shop/controllers/admin.js
@@ -2,16 +2,18 @@ const Product = require('../models/product');
 
 /** @type {import("express").RequestHandler} */
 exports.getAddProduct = (req, res, next) => {
-	req.res.render('admin/edit-product', {
+	if (!req.session.isAuthenticated) return res.redirect('/login');
+	res.render('admin/edit-product', {
 		pageTitle: 'Add Product',
 		path: '/admin/add-product',
 		editing: false,
-		isAuthenticated: req.isLoggedIn,
+		isAuthenticated: req.session.isAuthenticated,
 	});
 };
 
 /** @type {import("express").RequestHandler} */
 exports.postAddProduct = (req, res, next) => {
+	if (!req.session.isAuthenticated) return res.redirect('/login');
 	const title = req.body.title;
 	const imageUrl = req.body.imageUrl;
 	const price = parseFloat(req.body.price);
@@ -21,7 +23,7 @@ exports.postAddProduct = (req, res, next) => {
 		price,
 		description,
 		imageUrl,
-		user: req.user,
+		user: req.session.user,
 	});
 	product.save().then(product => {
 		console.log('Created Product');
@@ -31,6 +33,7 @@ exports.postAddProduct = (req, res, next) => {
 
 /** @type {import("express").RequestHandler} */
 exports.getEditProduct = (req, res, next) => {
+	if (!req.session.isAuthenticated) return res.redirect('/login');
 	const editMode = req.query.edit;
 	if (!editMode) {
 		return res.redirect('/');
@@ -43,7 +46,7 @@ exports.getEditProduct = (req, res, next) => {
 				path: '/admin/edit-product',
 				editing: editMode,
 				product,
-				isAuthenticated: req.isLoggedIn,
+				isAuthenticated: req.session.isAuthenticated,
 			});
 		})
 		.catch(err => console.log(err));
@@ -51,6 +54,7 @@ exports.getEditProduct = (req, res, next) => {
 
 /** @type {import("express").RequestHandler} */
 exports.postEditProduct = (req, res, next) => {
+	if (!req.session.isAuthenticated) return res.redirect('/login');
 	const prodId = req.body.productId;
 	const updatedTitle = req.body.title;
 	const updatedPrice = req.body.price;
@@ -72,13 +76,14 @@ exports.postEditProduct = (req, res, next) => {
 
 /** @type {import("express").RequestHandler} */
 exports.getProducts = (req, res, next) => {
+	if (!req.session.isAuthenticated) return res.redirect('/login');
 	Product.find()
 		.then(products => {
 			res.render('admin/products', {
 				prods: products,
 				pageTitle: 'Admin Products',
 				path: '/admin/products',
-				isAuthenticated: req.isLoggedIn,
+				isAuthenticated: req.session.isAuthenticated,
 			});
 		})
 		.catch(err => console.log(err));
@@ -86,6 +91,7 @@ exports.getProducts = (req, res, next) => {
 
 /** @type {import("express").RequestHandler} */
 exports.postDeleteProduct = (req, res, next) => {
+	if (!req.session.isAuthenticated) return res.redirect('/login');
 	const prodId = req.body.productId;
 	Product.findByIdAndDelete(prodId)
 		.then(result => {
